Move SoudEffectsPlayer implementation to the right file name
