+++ lib/alexandria/ui/acquire_dialog.rb
@@ -14,7 +14,7 @@ require "alexandria/scanners/keyboard"
 require "alexandria/ui/builder_base"
 require "alexandria/ui/barcode_animation"
 require "alexandria/ui/error_dialog"
-require "alexandria/ui/sound"
+require "alexandria/ui/sound_effects_player"
 
 module Alexandria
   module UI
